feat: add OG images for home, blog index, and privacy pages
Extract shared ImageResponse logic and generate opengraph-image routes for
site pages beyond blog posts so social previews include branded cards.
